Introduce null fultext provider
In order for the procedures to successfully load when the bloom extension
is disabled.
