+++ enterprise/fulltext-addon/src/main/java/org/neo4j/kernel/api/impl/fulltext/FulltextProvider.java
@@ -19,102 +19,82 @@
  */
 package org.neo4j.kernel.api.impl.fulltext;
 
-import java.io.File;
 import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
-import java.util.function.Consumer;
-import java.util.function.Function;
 
-import org.neo4j.graphdb.GraphDatabaseService;
-import org.neo4j.io.fs.FileSystemAbstraction;
-import org.neo4j.kernel.AvailabilityGuard;
 import org.neo4j.kernel.api.exceptions.InvalidArgumentsException;
 import org.neo4j.kernel.api.index.InternalIndexState;
-import org.neo4j.kernel.impl.transaction.log.TransactionIdStore;
-import org.neo4j.logging.Log;
-import org.neo4j.scheduler.JobScheduler;
 
-/**
- * Provider class that manages and provides fulltext indices. This is the main entry point for the fulltext addon.
- */
-public class FulltextProvider implements AutoCloseable
+public interface FulltextProvider extends AutoCloseable
 {
-    public static final String LUCENE_FULLTEXT_ADDON_PREFIX = "__lucene__fulltext__addon__";
-    public static final String FIELD_ENTITY_ID = LUCENE_FULLTEXT_ADDON_PREFIX + "internal__id__";
-
-    private final GraphDatabaseService db;
-    private final Log log;
-    private final TransactionIdStore transactionIdStore;
-    private final FulltextTransactionEventUpdater fulltextTransactionEventUpdater;
-    private final Set<String> nodeProperties;
-    private final Set<String> relationshipProperties;
-    private final Map<String,WritableFulltext> writableNodeIndices;
-    private final Map<String,WritableFulltext> writableRelationshipIndices;
-    private final FulltextUpdateApplier applier;
-    private final FulltextFactory factory;
-    private final ReadWriteLock configurationLock;
-
-    /**
-     * Creates a provider of fulltext indices for the given database. This is the entry point for all fulltext index
-     * operations.
-     *
-     * @param db Database that this provider should work with.
-     * @param log For logging errors.
-     * @param availabilityGuard Used for waiting with populating the index until the database is available.
-     * @param scheduler For background work.
-     * @param transactionIdStore Used for checking if the store has had transactions applied to it, while the fulltext
-     * @param fileSystem The filesystem to use.
-     * @param storeDir Store directory of the database.
-     * @param analyzerClassName The Lucene analyzer to use for the {@link LuceneFulltext} created by this factory.
-     */
-    public FulltextProvider( GraphDatabaseService db, Log log, AvailabilityGuard availabilityGuard,
-                             JobScheduler scheduler, TransactionIdStore transactionIdStore,
-                             FileSystemAbstraction fileSystem, File storeDir,
-                             String analyzerClassName ) throws IOException
+    FulltextProvider NULL_PROVIDER = new FulltextProvider()
     {
-        this.db = db;
-        this.log = log;
-        this.transactionIdStore = transactionIdStore;
-        applier = new FulltextUpdateApplier( log, availabilityGuard, scheduler );
-        applier.start();
-        factory = new FulltextFactory( fileSystem, storeDir, analyzerClassName );
-        fulltextTransactionEventUpdater = new FulltextTransactionEventUpdater( this, log, applier );
-        nodeProperties = ConcurrentHashMap.newKeySet();
-        relationshipProperties = ConcurrentHashMap.newKeySet();
-        writableNodeIndices = new ConcurrentHashMap<>();
-        writableRelationshipIndices = new ConcurrentHashMap<>();
-        configurationLock = new ReentrantReadWriteLock( true );
-    }
-
-    public void registerTransactionEventHandler() throws IOException
-    {
-        db.registerTransactionEventHandler( fulltextTransactionEventUpdater );
-    }
-
-    private boolean matchesConfiguration( WritableFulltext index ) throws IOException
-    {
-        long txId = transactionIdStore.getLastCommittedTransactionId();
-        FulltextIndexConfiguration currentConfig =
-                new FulltextIndexConfiguration( index.getAnalyzerName(), index.getProperties(), txId );
-
-        FulltextIndexConfiguration storedConfig;
-        try ( ReadOnlyFulltext indexReader = index.getIndexReader() )
+        @Override
+        public void registerTransactionEventHandler() throws IOException
         {
-            storedConfig = indexReader.getConfigurationDocument();
+            throw noProvider();
         }
-        return storedConfig == null && index.getProperties().isEmpty() ||
-               storedConfig != null && storedConfig.equals( currentConfig );
-    }
+
+        @Override
+        public void awaitPopulation()
+        {
+            throw noProvider();
+        }
+
+        @Override
+        public void openIndex( String identifier, FulltextIndexType type ) throws IOException
+        {
+            throw noProvider();
+        }
+
+        @Override
+        public void createIndex( String identifier, FulltextIndexType type, List<String> properties ) throws IOException
+        {
+            throw noProvider();
+        }
+
+        @Override
+        public ReadOnlyFulltext getReader( String identifier, FulltextIndexType type ) throws IOException
+        {
+            throw noProvider();
+        }
+
+        @Override
+        public Set<String> getProperties( String identifier, FulltextIndexType type )
+        {
+            throw noProvider();
+        }
+
+        @Override
+        public InternalIndexState getState( String identifier, FulltextIndexType type )
+        {
+            throw noProvider();
+        }
+
+        @Override
+        public void changeIndexedProperties( String identifier, FulltextIndexType type, List<String> propertyKeys )
+                throws IOException, InvalidArgumentsException
+        {
+            throw noProvider();
+        }
+
+        @Override
+        public void close() throws Exception
+        {
+            throw noProvider();
+        }
+
+        private RuntimeException noProvider()
+        {
+            return new UnsupportedOperationException(
+                    "There is no fulltext provider for this database. Make sure that the feature you are tyring to use is enabled" );
+        }
+    };
+    String LUCENE_FULLTEXT_ADDON_PREFIX = "__lucene__fulltext__addon__";
+    String FIELD_ENTITY_ID = LUCENE_FULLTEXT_ADDON_PREFIX + "internal__id__";
+
+    void registerTransactionEventHandler() throws IOException;
 
     /**
      * Wait for the asynchronous background population, if one is on-going, to complete.
@@ -124,96 +104,11 @@ public class FulltextProvider implements AutoCloseable
      *
      * @throws RuntimeException If it was not possible to wait for the population to finish, for some reason.
      */
-    public void awaitPopulation()
-    {
-        try
-        {
-            applier.writeBarrier().awaitCompletion();
-        }
-        catch ( ExecutionException e )
-        {
-            throw new AssertionError( "The writeBarrier operation should never throw an exception", e );
-        }
-        catch ( IOException e )
-        {
-            throw new UncheckedIOException( e );
-        }
-    }
+    void awaitPopulation();
 
-    public void openIndex( String identifier, FulltextIndexType type ) throws IOException
-    {
-        LuceneFulltext index = factory.openFulltextIndex( identifier, type );
-        register( index );
-    }
+    void openIndex( String identifier, FulltextIndexType type ) throws IOException;
 
-    public void createIndex( String identifier, FulltextIndexType type, List<String> properties )
-            throws IOException
-    {
-        LuceneFulltext index = factory.createFulltextIndex( identifier, type, properties );
-        register( index );
-    }
-
-    private void register( LuceneFulltext fulltextIndex ) throws IOException
-    {
-        configurationLock.writeLock().lock();
-        try
-        {
-            WritableFulltext writableFulltext = new WritableFulltext( fulltextIndex );
-            writableFulltext.open();
-            if ( fulltextIndex.getType() == FulltextIndexType.NODES )
-            {
-                if ( !matchesConfiguration( writableFulltext ) )
-                {
-                    writableFulltext.drop();
-                    writableFulltext.open();
-                    if ( !writableFulltext.getProperties().isEmpty() )
-                    {
-                        applier.populateNodes( writableFulltext, db );
-                    }
-                }
-                writableNodeIndices.put( fulltextIndex.getIdentifier(), writableFulltext );
-                nodeProperties.addAll( fulltextIndex.getProperties() );
-            }
-            else
-            {
-                if ( !matchesConfiguration( writableFulltext ) )
-                {
-                    writableFulltext.drop();
-                    writableFulltext.open();
-                    if ( !writableFulltext.getProperties().isEmpty() )
-                    {
-                        applier.populateRelationships( writableFulltext, db );
-                    }
-                }
-                writableRelationshipIndices.put( fulltextIndex.getIdentifier(), writableFulltext );
-                relationshipProperties.addAll( fulltextIndex.getProperties() );
-            }
-        }
-        finally
-        {
-            configurationLock.writeLock().unlock();
-        }
-    }
-
-    String[] getNodeProperties()
-    {
-        return nodeProperties.toArray( new String[0] );
-    }
-
-    String[] getRelationshipProperties()
-    {
-        return relationshipProperties.toArray( new String[0] );
-    }
-
-    Collection<WritableFulltext> writableNodeIndices()
-    {
-        return Collections.unmodifiableCollection( writableNodeIndices.values() );
-    }
-
-    Collection<WritableFulltext> writableRelationshipIndices()
-    {
-        return Collections.unmodifiableCollection( writableRelationshipIndices.values() );
-    }
+    void createIndex( String identifier, FulltextIndexType type, List<String> properties ) throws IOException;
 
     /**
      * Returns a reader for the specified index.
@@ -223,137 +118,11 @@ public class FulltextProvider implements AutoCloseable
      * @return A {@link ReadOnlyFulltext} for the index, or null if no such index is found.
      * @throws IOException
      */
-    public ReadOnlyFulltext getReader( String identifier, FulltextIndexType type ) throws IOException
-    {
-        WritableFulltext writableFulltext = getIndexMap( type ).get( identifier );
-        if ( writableFulltext == null )
-        {
-            throw new IllegalArgumentException( "No such " + type + " index '" + identifier + "'." );
-        }
-        return writableFulltext.getIndexReader();
-    }
+    ReadOnlyFulltext getReader( String identifier, FulltextIndexType type ) throws IOException;
 
-    private Map<String,WritableFulltext> getIndexMap( FulltextIndexType type )
-    {
-        switch ( type )
-        {
-        case NODES:
-            return writableNodeIndices;
-        case RELATIONSHIPS:
-            return writableRelationshipIndices;
-        default:
-            throw new IllegalArgumentException( "No such fulltext index type: " + type );
-        }
-    }
+    Set<String> getProperties( String identifier, FulltextIndexType type );
 
-    public Set<String> getProperties( String identifier, FulltextIndexType type )
-    {
-        return applyToMatchingIndex( identifier, type, WritableFulltext::getProperties );
-    }
+    InternalIndexState getState( String identifier, FulltextIndexType type );
 
-    private <E> E applyToMatchingIndex(
-            String identifier, FulltextIndexType type, Function<WritableFulltext,E> function )
-    {
-        if ( type == FulltextIndexType.NODES )
-        {
-            return function.apply( writableNodeIndices.get( identifier ) );
-        }
-        else
-        {
-            return function.apply( writableRelationshipIndices.get( identifier ) );
-        }
-    }
-
-    public InternalIndexState getState( String identifier, FulltextIndexType type )
-    {
-        return applyToMatchingIndex( identifier, type, WritableFulltext::getState );
-    }
-
-    void drop( String identifier, FulltextIndexType type ) throws IOException
-    {
-        configurationLock.writeLock().lock();
-        try
-        {
-            // Wait for the queue of updates to drain, before deleting an index.
-            awaitPopulation();
-            if ( type == FulltextIndexType.NODES )
-            {
-                writableNodeIndices.remove( identifier ).drop();
-            }
-            else
-            {
-                writableRelationshipIndices.remove( identifier ).drop();
-            }
-            rebuildProperties();
-        }
-        finally
-        {
-            configurationLock.writeLock().unlock();
-        }
-    }
-
-    private void rebuildProperties()
-    {
-        nodeProperties.clear();
-        relationshipProperties.clear();
-        writableNodeIndices.forEach( ( s, index ) -> nodeProperties.addAll( index.getProperties() ) );
-        writableRelationshipIndices.forEach( ( s, index ) -> relationshipProperties.addAll( index.getProperties() ) );
-    }
-
-    Lock readLockIndexConfiguration()
-    {
-        Lock lock = configurationLock.readLock();
-        lock.lock();
-        return lock;
-    }
-
-    public void changeIndexedProperties( String identifier, FulltextIndexType type, List<String> propertyKeys )
-            throws IOException, InvalidArgumentsException
-    {
-        configurationLock.writeLock().lock();
-        try
-        {
-            if ( propertyKeys.stream().anyMatch( s -> s.startsWith( FulltextProvider.LUCENE_FULLTEXT_ADDON_PREFIX ) ) )
-            {
-                throw new InvalidArgumentsException(
-                        "It is not possible to index property keys starting with " +
-                        FulltextProvider.LUCENE_FULLTEXT_ADDON_PREFIX );
-            }
-            Set<String> currentProperties = getProperties( identifier, type );
-            if ( !currentProperties.containsAll( propertyKeys ) || !propertyKeys.containsAll( currentProperties ) )
-            {
-                drop( identifier, type );
-                createIndex( identifier, type, propertyKeys );
-            }
-        }
-        finally
-        {
-            configurationLock.writeLock().unlock();
-        }
-    }
-
-    /**
-     * Closes the provider and all associated resources.
-     */
-    @Override
-    public void close()
-    {
-        db.unregisterTransactionEventHandler( fulltextTransactionEventUpdater );
-        applier.stop();
-        Consumer<WritableFulltext> fulltextCloser = luceneFulltextIndex ->
-        {
-            try
-            {
-                luceneFulltextIndex.saveConfiguration( transactionIdStore.getLastCommittedTransactionId() );
-                luceneFulltextIndex.close();
-            }
-            catch ( IOException e )
-            {
-                log.error( "Unable to close fulltext index.", e );
-            }
-        };
-        writableNodeIndices.values().forEach( fulltextCloser );
-        writableRelationshipIndices.values().forEach( fulltextCloser );
-    }
-
+    void changeIndexedProperties( String identifier, FulltextIndexType type, List<String> propertyKeys ) throws IOException, InvalidArgumentsException;
 }
