+++ enterprise/fulltext-addon/src/test/java/org/neo4j/kernel/api/impl/fulltext/LuceneFulltextUpdaterTest.java
@@ -722,7 +722,7 @@ public class LuceneFulltextUpdaterTest extends LuceneFulltextTestSupport
     @Test
     public void shouldBeAbleToDropAndReaddIndex() throws Exception
     {
-        try ( FulltextProvider provider = createProvider() )
+        try ( FulltextProviderImpl provider = createProvider() )
         {
             provider.createIndex( "nodes", NODES, singletonList( "prop" ) );
             provider.registerTransactionEventHandler();
